feat: Improve human AI
- The human AI now has a field of view and uses a state machine.
- The human starts in a "patrol state" where it simply moves randomly to certain locations.
- Once the player enters the FOV, the human moves to a "chase state" where it chases the player.
- After the player gets out of the view radius, the human will still try to chase the player for
a little while before giving up and going back to its "patrol state". If the player gets back
into the FOV, then the human returns to its "chase state".
- The human has another smaller "detection" radius that will send the human to a "suspicious state"
whenever the player is in that zone regardless of being in the FOV.
- In the "suspicious state" the human rotates in place for some time before returning to the "patrol state"
unless the player is detected in this state. In that case, the human will then transisiton to the "chase state".
- Whenever the player is in the attack range in the FOV, the human transitions to the "attack state" which has not
been fully implemented yet.