fix: AI detection and patroling
The game will only end when there are in attack range of the player, not the trailing food.
Humans will track the player when they detect the trailing food instead of tracking the trailing food.
Fixed the issue where the humans will sometimes group together and stop moving until the player gets in range.
Apparently, the humans somehow all had the same patrol points. So I hade to make sure each human was not sharing
a patrol point with another.