feat: Human Model + Animations & Idle State
Imported the CubePeopleDemo assets from the Unity Asset Store. The HumanManager Script
was incorporated to an actual human model and updated the script to support animations.
The animations change based on the current state. Idle and Suspicious State have the idle
animation. Patrol State has the wlaking animation. The Chase State has the running animation.
I also made a proper idle state and refactored code that was related to idling into the IdleState
script.