feat: More AI improvement; import character model
Adjusted the attack range to be much smaller than the detection radius.
Increased overall speed of the AI since it was a little bit too slow earlier.
Imported the main character model (apple with big eyeballs) to the scene to test with the AI.
Everything still works, just had to make a minor adjustement to the PlayerMovement
script so that model will face the direction the player is traveling.